test: move change history test code to it's own file
